@@ -12,41 +12,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.1.3",
-    "@angular/common": "^4.1.3",
-    "@angular/compiler": "^4.1.3",
-    "@angular/core": "^4.1.3",
-    "@angular/forms": "^4.1.3",
-    "@angular/http": "^4.1.3",
-    "@angular/platform-browser": "^4.1.3",
-    "@angular/platform-browser-dynamic": "^4.1.3",
-    "@angular/router": "^4.1.3",
-    "@ng-bootstrap/ng-bootstrap": "latest",
+    "@angular/animations": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
     "core-js": "^2.4.1",
-    "devextreme": "^17.1.3",
-    "devextreme-angular": "^17.1.3",
+    "devextreme": "^17.2.7",
+    "devextreme-angular": "^17.2.7",
     "font-awesome": "^4.7.0",
-    "primeng": "^4.0.1",
-    "rxjs": "^5.1.0",
+    "jasmine": "^3.1.0",
+    "jszip": "^3.1.5",
+    "rxjs": "^6.1.0",
     "zone.js": "^0.8.11"
   },
   "devDependencies": {
-    "@angular/cli": "^1.0.4",
-    "@angular/compiler-cli": "^4.1.3",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
     "@types/jasmine": "^2.5.47",
-    "@types/node": "~7.0.21",
-    "codelyzer": "~3.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "@types/node": "^10.0.4",
+    "codelyzer": "^4.3.0",
+    "jasmine-core": "^3.1.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^2.0.2",
+    "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "~0.2.2",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.2.0",
-    "typescript": "~2.3.2"
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^1.0.0",
+    "protractor": "^5.3.1",
+    "ts-node": "^6.0.3",
+    "tslint": "^5.10.0",
+    "typescript": "^2.7.2"
   }
 }
