@@ -4,7 +4,7 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 import { Router, NavigationStart, NavigationEnd, RouterModule } from '@angular/router';
-import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { routing } from './app.routing';
 
@@ -36,7 +36,7 @@ declare var NProgress: any;
         HttpModule,
         RouterModule,
         StructureModule,
-        FeaturesModule,
+        FeaturesModule,        
         NgbModule.forRoot(),
         routing
     ],
