@@ -3,10 +3,11 @@ import { NgModule }      from '@angular/core';
 import { CommonModule }  from '@angular/common';
 import { Routes, RouterModule }  from '@angular/router';
 
-import {DataTableModule,SharedModule,DialogModule,ButtonModule} from 'primeng/primeng';
+import { DxDateBoxModule } from 'devextreme-angular'
 
 import { ConsultationList } from './consultation.list'
 import { ConsultationAdd } from './consultation.add'
+import { PatientFormComponent } from '../../components/patient-form/patient-form.component';
 
 export const routes: Routes = [
   { path: 'consultation/list', component: ConsultationList },
@@ -16,17 +17,15 @@ export const routes: Routes = [
 @NgModule({
   imports: [    
     CommonModule,
-    DataTableModule,
-    DialogModule,
-    ButtonModule,
-    SharedModule,
     FormsModule,
+    DxDateBoxModule,
     ReactiveFormsModule,
     RouterModule.forChild(routes)
   ],
   declarations: [
     ConsultationList,
-    ConsultationAdd
+    ConsultationAdd,
+    PatientFormComponent
   ]
 
 })
